Reconfigure ALE and other shortcuts in vim
Added the :ALEStopAllLSPs one because Python LSP stuff was giving me
trouble, and the most effective way to work around for now seems to be
to stop the LSP with this command and then save the buffer.

Added a binding for :ALERename as well. After renaming, you have to save
all unsaved buffers for things to actually change on disk.

Added a shortcut for :FZF because omnisearch needs to be as ergonomic as
possible

Added commands to save/load a session, so that when I do feel the need
to quit vim for whatever reason, I have quick muscle memory to restore
my session